moved pagination as general ui component
